Refactor 'SourceBridge' into SourceEmbedded and PluginBridge
We want to re-use the Source Bridge protocol for other iframe embedded purposes,
notably embedded Connect applications inside customer applications. The plan is
to:
a) Extract a general 'embedded' client that doesn't have anything plugin-
   specific.
b) Maintain a very similar API to the current SourceBridge as 'PluginBridge' ,
   that uses the general embedded client.
c) Add a new 'Embedder' client for the parent applications.

This PR is (a) and (b).

Changes:
- Drop 'MessageType' string union since the generic client does not know about
  all of the different events applications might send (we will probably build
  that enum into the higher-level SDK that can embed e.g. a form iframe. We just
  use 'string' for message type fields at this level.
- Rename 'SourceBridge' to 'SourceEmbedded' and remove the plugin-specific API.
- Implement the old SourceBridge plugin API as PluginBridge.
- Move plugin-specific message interfaces to plugins/Messages.ts
- Rename existing tests to legacy_plugin
- Copy-paste plugin tests so that we are testing the new-style usage of
  PluginBridge as well as the backwards-compatible way.
- Update README
- Add CHANGELOG.md