Introduce BridgeHost, shift auth to request/response, SourceEmbedded=>BridgeGuest
Introduce BridgeHost, a class to manage the host-application side of the Source
Bridge protocol. In this PR we also switch the underlying bridge protocol to
exchange authentication tokens using a request/response flow initiated by the
guest, rather than having the host push tokens down to the guest as events. This
does not change the API of PluginBridge, but it allows for more flexibility on
the host side including retrieving a new token from a backend if required.

- Rename 'SourceEmbedded' to 'BridgeGuest' (the guest side of the bridge).
- Introduce BridgeHost, a class to manage the host side of the bridge, based on
  code within the Source clinical app that manages plugin guests.
- Switch the authentication mechanism to request/response rather than expecting
  the parent app to push credentials as an event. This has no impact on the
  PluginBridge API.
- Define all the interfaces we use to exchange data over the bridge
  (enevelopes, requests, responses, events; and the 'hello', 'authentication',
  and 'ready' messages.)
- Extract some helpers from test code.
- Update README and CHANGELOG