[PluginBridge] Remove global singleton
We had previously defined SourceBridge as a singleton in the bridge library.
Trying to use our new PluginBridge the same way (at least, via yarn link in the
demo plugin project) was leading to a `PluginBridge is undefined` error.

Making it a non-global/non-singleton seems to make it work. It just means a few
more changes for plugin developers.

Note: I don’t fully understand why this would be - is it due to actual code
changes or is it only a problem because of referencing a local repo via 'link'?
Because we need to test that our advice for upgrading plugins will work before
we publish v0.1 of Bridge, I think it's prudent to make this change.

- Remove global singleton and export the class instead (as PluginBridge)
- Update changelog and advice for updating existing plugins.
- Update e2e tests