[Build] Remove 'node' build
When using this library in a Next.js project via `yarn link`, it worked at build
time but was being reported as not defined at run time. Removing the node build
and making the `main` property in package.json point to the browser bundle seems
to fix this.

The build was initially based on our config of the source-client which does need
to be built for browsers and node. This library is browser-only.