[PluginBridge] Remove global singletons (#8)
We had previously defined SourceBridge as a singleton in the bridge library.
Trying to use our new PluginBridge the same way (at least, via yarn link in the
demo plugin project) was leading to a `PluginBridge is undefined` error.

Making PluginBridge a non-global/non-singleton seems to make it work. It just
means a few more changes for plugin developers.

Similarly, when using this library in a Next.js project, the parent global is not
defined in the build-time environment (I think it's because Next code can run
server-side as well as in the browser?).

When using this library in a Next.js project via `yarn link`, it worked at build
time but was being reported as not defined at run time. Removing the node build
and making the `main` property in package.json point to the browser bundle seems
to fix this. The build was initially based on our config of the source-client which
does need to be built for browsers and node. This library is browser-only.


- Remove global PluginBridge singleton and export the class instead (as PluginBridge)
- Remove the global BridgeGuest singleton and have clients instantiate a
  BridgeGuest() class instead.
- Update changelog and advice for updating existing plugins.
- Update e2e tests
- Remove node build
- Wire debugging into BridgeGuest and PluginBridge and tweak debug() methods